refactor(chat): split finalizers and local bridge
