+++ package.json
@@ -42,6 +42,7 @@
     "test:release:smoke": "node scripts/run-release-regression.mjs --mode all --level smoke",
     "test:release:nightly": "node scripts/run-release-regression.mjs --mode all --level nightly",
     "test:release:ui": "node --import tsx scripts/run-electron-ui-smoke.ts",
+    "release:gate": "node scripts/release-gate.mjs",
     "release:preflight": "npm test && npm run typecheck && npm run typecheck:runtime && npm run build:server && npm run build:main && npm run build:renderer && npm run test:release && npm run test:release:ui",
     "build:server": "node scripts/build-server.mjs",
     "build:server:win": "node scripts/build-server.mjs win32 x64",
