+++ server/chat/provider-meta.ts
@@ -1,0 +1,1 @@
+export * from "./provider-route-meta.js";
